app/requirements.txt
beautifulsoup4
lxml
requests
python-dotenv
numpy
scipy